[v11] Simplify Component System Usage
